updated the paths to work universally

+++ setupWindowsEnv.ps1
@@ -1,5 +1,6 @@
 # Set execution policy for the session
 Set-ExecutionPolicy Bypass -Scope Process -Force
+Set-ExecutionPolicy RemoteSigned
 
 # Install package managers
 [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072
@@ -7,7 +8,7 @@ Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://com
 Invoke-RestMethod -Uri https://get.scoop.sh | Invoke-Expression
 
 # Install packages using Chocolatey
-$packages = 'powershell-core', 'git', 'gh', 'vscode', 'nodejs.install', 'fzf', 'neovim', 'ripgrep', 
+$packages = 'winget', 'powershell-core', 'git', 'gh', 'vscode', 'nodejs.install', 'fzf', 'neovim', 'ripgrep', 
 'lazygit', 'bat', 'fd', 'mingw', 'cmake', 'llvm', 'python', 'pyenv-win', 'nerd-fonts-JetBrainsMono', 
 'nerd-fonts-Hack'
 foreach ($package in $packages)
@@ -15,12 +16,15 @@ foreach ($package in $packages)
   choco install $package -y
 }
 
+# Install NuGet provider for Powershell
+Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser
+
 # Install packages using winget
 $wingetPackages = 'Microsoft.WindowsTerminal', 'Microsoft.PowerToys', 'JanDeDobbeleer.OhMyPosh', 
 'Google.Chrome', 'Mozilla.Firefox'
 foreach ($app in $wingetPackages)
 {
-  winget install -e --id=$app
+  winget install -e --id=$app -a
 }
 
 # Install PowerShell modules
@@ -34,23 +38,31 @@ foreach ($module in $modules)
 git clone https://github.com/LazyVim/starter $env:LOCALAPPDATA\nvim
 Remove-Item $env:LOCALAPPDATA\nvim\.git -Recurse -Force
 
-# Configuration settings for Powershell and Terminal
-$configSource = "~\Downloads\config_files"
+git clone https://github.com/nxstynate/config_files ~\Downloads
+
+# Configuration settings for PowerShell and Terminal
+$configSource = "$env:USERPROFILE\Downloads\config_files"
 $nvimConfigSource = "$configSource\neovim\nvim_devaslife_2024\*"
-$powerShellConfigSource = "$configSource\windows\PowerShell\"
+$powerShellConfigSource = "$configSource\windows\PowerShell\*"
 $terminalConfigSource = "$configSource\windows\Terminal\settings.json"
-$powerToysConfigSource = "$configSource\windows\PowerToys\"
+$powerToysConfigSource = "$configSource\windows\PowerToys\*"
 
+# Copy items to their respective locations
 Copy-Item $nvimConfigSource "$env:LOCALAPPDATA\nvim\." -Force -Recurse
-Copy-Item "$powerShellConfigSource\*" "~\Documents\PowerShell\" -Force -Recurse
-Copy-Item $terminalConfigSource "~\AppData\Local\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\LocalState\settings.json" -Force -Recurse
-Copy-Item "$powerToysConfigSource\*" "~\AppData\Local\Microsoft\PowerToys\" -Force -Recurse
+Copy-Item $powerShellConfigSource "$env:USERPROFILE\Documents\PowerShell\" -Force -Recurse
+Copy-Item $terminalConfigSource "$env:LOCALAPPDATA\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\LocalState\settings.json" -Force -Recurse
+Copy-Item $powerToysConfigSource "$env:LOCALAPPDATA\Microsoft\PowerToys\" -Force -Recurse
 
-Remove-Item $configSource -Recurse -Force
-
-# Restart PowerToys to apply new settings
-Stop-Process -Name "PowerToys" -Force
-Start-Process -FilePath "C:\Program Files\PowerToys\PowerToys.exe"
+# Stop and start PowerToys (if installed and running)
+$powerToysProcess = Get-Process -Name "PowerToys" -ErrorAction SilentlyContinue
+if ($powerToysProcess)
+{
+  Stop-Process -Name "PowerToys" -Force
+  Start-Process -FilePath "C:\Program Files\PowerToys\PowerToys.exe"
+} else
+{
+  Write-Output "PowerToys is not running. Skipping restart."
+}
 
 # Node package managers
 npm install -g npm yarn
@@ -74,6 +86,15 @@ pip install --user pipenv
 # }
 #
 #
+
+
+
+
+
+
+
+
+
